Removed Advertiser, create/store Item, cleanup
Removed Advertiser to make a single type of user. Attributes belonging to Advertiser has been moved to User. Added ability to create and store a new item, linked to the user. Cleaned up the images used in testing. Added a minimum bid to items. Added a times-viewed attribute to items.