Changed behaviour of marking items sold
When item gets marked as sold, it sends a notification to the users who bid on it. The buyer gets a different message than the others who bid on it. Small changes to the database as well as adding a table for notifs. Amount of items sold now no longer tallied on user account, instead searched through Count in database.