Payment processing, promoting
Adds processing of payment without actual payment processing. You can pay an existing invoice or promote an item and pay it on the spot. When item is not paid for directly, the invoice remains. Promoted items appear on the top of the list, ordered by last promoted item. Further cleaning and commenting.